+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.0.0]
 
 - Replace ObsPy usage with pymseed for much smaller requirements.
+- Relative paths in the server config file are now relative to the config file directory.
 - Add IPv6 support.
 - Fix overrun of fields in /summary endpoint output with high resolution time stamps.
 - Significant refactoring into simpler a code base.